Graphics and Stability Update
22/10/2024:

added adaptive texture streaming and changed the graphics settings in general to help support lower end PC's that we are targeting for our audience. Testing before this update showed (in editor) 2gb of ram and ~30%GPU usage of a high end graphics card. This update should reduce some of that load and help with lower end PC's